Refactor models to DataBase.Entities namespace
Moved User, Post, and Following models to the SocialNetwork.Models.DataBase.Entities namespace and updated references in SocialNetworkContext and Program.cs. This improves project structure and organization for entity classes.
