+++ server/SocialNetwork/SocialNetwork/Models/DataBase/Entities/Post.cs
@@ -1,4 +1,4 @@
-﻿namespace SocialNetwork.Models;
+﻿namespace SocialNetwork.Models.DataBase.Entities;
 public class Post {
     public long Id { get; set; }
     public required long UserId { get; set; }
